requirements.txt
slack_bolt>=1.19.0
python-dotenv
aiohttp
openai